@@ -45,7 +45,7 @@ public:
 
     this->timer_ = this->create_wall_timer(
       std::chrono::milliseconds(500),
-      std::bind(&FibonacciActionClient::send_goal, this));
+      [this]() {return this->send_goal();});
   }
 
   ACTION_TUTORIALS_CPP_PUBLIC
@@ -67,66 +67,59 @@ public:
     RCLCPP_INFO(this->get_logger(), "Sending goal");
 
     auto send_goal_options = rclcpp_action::Client<Fibonacci>::SendGoalOptions();
-    send_goal_options.goal_response_callback =
-      std::bind(&FibonacciActionClient::goal_response_callback, this, _1);
-    send_goal_options.feedback_callback =
-      std::bind(&FibonacciActionClient::feedback_callback, this, _1, _2);
-    send_goal_options.result_callback =
-      std::bind(&FibonacciActionClient::result_callback, this, _1);
+    send_goal_options.goal_response_callback = [this](
+      const GoalHandleFibonacci::SharedPtr & goal_handle)
+      {
+        if (!goal_handle) {
+          RCLCPP_ERROR(this->get_logger(), "Goal was rejected by server");
+        } else {
+          RCLCPP_INFO(this->get_logger(), "Goal accepted by server, waiting for result");
+        }
+      };
+
+    send_goal_options.feedback_callback = [this](
+      GoalHandleFibonacci::SharedPtr,
+      const std::shared_ptr<const Fibonacci::Feedback> feedback)
+      {
+        std::stringstream ss;
+        ss << "Next number in sequence received: ";
+        for (auto number : feedback->partial_sequence) {
+          ss << number << " ";
+        }
+        RCLCPP_INFO(this->get_logger(), "%s", ss.str().c_str());
+      };
+
+    send_goal_options.result_callback = [this](
+      const GoalHandleFibonacci::WrappedResult & result)
+      {
+        switch (result.code) {
+          case rclcpp_action::ResultCode::SUCCEEDED:
+            break;
+          case rclcpp_action::ResultCode::ABORTED:
+            RCLCPP_ERROR(this->get_logger(), "Goal was aborted");
+            return;
+          case rclcpp_action::ResultCode::CANCELED:
+            RCLCPP_ERROR(this->get_logger(), "Goal was canceled");
+            return;
+          default:
+            RCLCPP_ERROR(this->get_logger(), "Unknown result code");
+            return;
+        }
+        std::stringstream ss;
+        ss << "Result received: ";
+        for (auto number : result.result->sequence) {
+          ss << number << " ";
+        }
+        RCLCPP_INFO(this->get_logger(), "%s", ss.str().c_str());
+        rclcpp::shutdown();
+      };
+
     this->client_ptr_->async_send_goal(goal_msg, send_goal_options);
   }
 
 private:
   rclcpp_action::Client<Fibonacci>::SharedPtr client_ptr_;
   rclcpp::TimerBase::SharedPtr timer_;
-
-  ACTION_TUTORIALS_CPP_LOCAL
-  void goal_response_callback(GoalHandleFibonacci::SharedPtr goal_handle)
-  {
-    if (!goal_handle) {
-      RCLCPP_ERROR(this->get_logger(), "Goal was rejected by server");
-    } else {
-      RCLCPP_INFO(this->get_logger(), "Goal accepted by server, waiting for result");
-    }
-  }
-
-  ACTION_TUTORIALS_CPP_LOCAL
-  void feedback_callback(
-    GoalHandleFibonacci::SharedPtr,
-    const std::shared_ptr<const Fibonacci::Feedback> feedback)
-  {
-    std::stringstream ss;
-    ss << "Next number in sequence received: ";
-    for (auto number : feedback->partial_sequence) {
-      ss << number << " ";
-    }
-    RCLCPP_INFO(this->get_logger(), "%s", ss.str().c_str());
-  }
-
-  ACTION_TUTORIALS_CPP_LOCAL
-  void result_callback(const GoalHandleFibonacci::WrappedResult & result)
-  {
-    switch (result.code) {
-      case rclcpp_action::ResultCode::SUCCEEDED:
-        break;
-      case rclcpp_action::ResultCode::ABORTED:
-        RCLCPP_ERROR(this->get_logger(), "Goal was aborted");
-        return;
-      case rclcpp_action::ResultCode::CANCELED:
-        RCLCPP_ERROR(this->get_logger(), "Goal was canceled");
-        return;
-      default:
-        RCLCPP_ERROR(this->get_logger(), "Unknown result code");
-        return;
-    }
-    std::stringstream ss;
-    ss << "Result received: ";
-    for (auto number : result.result->sequence) {
-      ss << number << " ";
-    }
-    RCLCPP_INFO(this->get_logger(), "%s", ss.str().c_str());
-    rclcpp::shutdown();
-  }
 };  // class FibonacciActionClient
 
 }  // namespace action_tutorials_cpp
