@@ -33,7 +33,7 @@ Client::Client(const rclcpp::NodeOptions & options)
   // Note(dhood): The timer period must be greater than the duration of the timer callback.
   // Otherwise, the timer can starve a single-threaded executor.
   // See https://github.com/ros2/rclcpp/issues/392 for updates.
-  timer_ = create_wall_timer(2s, std::bind(&Client::on_timer, this));
+  timer_ = create_wall_timer(2s, [this]() {return this->on_timer();});
 }
 
 void Client::on_timer()
