@@ -14,6 +14,7 @@
 
 #include <memory>
 #include <string>
+#include <utility>
 #include <vector>
 
 #include "class_loader/class_loader.hpp"
@@ -35,7 +36,7 @@ int main(int argc, char * argv[])
   rclcpp::Logger logger = rclcpp::get_logger(DLOPEN_COMPOSITION_LOGGER_NAME);
   rclcpp::executors::SingleThreadedExecutor exec;
   rclcpp::NodeOptions options;
-  std::vector<class_loader::ClassLoader *> loaders;
+  std::vector<std::unique_ptr<class_loader::ClassLoader>> loaders;
   std::vector<rclcpp_components::NodeInstanceWrapper> node_wrappers;
 
   std::vector<std::string> libraries;
@@ -44,7 +45,7 @@ int main(int argc, char * argv[])
   }
   for (auto library : libraries) {
     RCLCPP_INFO(logger, "Load library %s", library.c_str());
-    auto loader = new class_loader::ClassLoader(library);
+    auto loader = std::make_unique<class_loader::ClassLoader>(library);
     auto classes = loader->getAvailableClasses<rclcpp_components::NodeFactory>();
     for (auto clazz : classes) {
       RCLCPP_INFO(logger, "Instantiate class %s", clazz.c_str());
@@ -54,7 +55,7 @@ int main(int argc, char * argv[])
       node_wrappers.push_back(wrapper);
       exec.add_node(node);
     }
-    loaders.push_back(loader);
+    loaders.push_back(std::move(loader));
   }
 
   exec.spin();
