@@ -13,7 +13,6 @@
 // limitations under the License.
 
 #include <chrono>
-#include <iostream>
 #include <memory>
 #include <string>
 #include <thread>
@@ -26,8 +25,6 @@
 
 #include "rclcpp_lifecycle/lifecycle_node.hpp"
 #include "rclcpp_lifecycle/lifecycle_publisher.hpp"
-
-#include "rcutils/logging_macros.h"
 
 #include "std_msgs/msg/string.hpp"
 
@@ -119,7 +116,7 @@ public:
     // available.
     pub_ = this->create_publisher<std_msgs::msg::String>("lifecycle_chatter", 10);
     timer_ = this->create_wall_timer(
-      1s, std::bind(&LifecycleTalker::publish, this));
+      1s, [this]() {return this->publish();});
 
     RCLCPP_INFO(get_logger(), "on_configure() is called.");
 
@@ -152,7 +149,7 @@ public:
     // Overriding this method is optional, a lot of times the default is enough.
     LifecycleNode::on_activate(state);
 
-    RCUTILS_LOG_INFO_NAMED(get_name(), "on_activate() is called.");
+    RCLCPP_INFO(get_logger(), "on_activate() is called.");
 
     // Let's sleep for 2 seconds.
     // We emulate we are doing important
@@ -188,7 +185,7 @@ public:
     // Overriding this method is optional, a lot of times the default is enough.
     LifecycleNode::on_deactivate(state);
 
-    RCUTILS_LOG_INFO_NAMED(get_name(), "on_deactivate() is called.");
+    RCLCPP_INFO(get_logger(), "on_deactivate() is called.");
 
     // We return a success and hence invoke the transition to the next
     // step: "inactive".
@@ -219,7 +216,7 @@ public:
     timer_.reset();
     pub_.reset();
 
-    RCUTILS_LOG_INFO_NAMED(get_name(), "on cleanup is called.");
+    RCLCPP_INFO(get_logger(), "on cleanup is called.");
 
     // We return a success and hence invoke the transition to the next
     // step: "unconfigured".
@@ -250,10 +247,7 @@ public:
     timer_.reset();
     pub_.reset();
 
-    RCUTILS_LOG_INFO_NAMED(
-      get_name(),
-      "on shutdown is called from state %s.",
-      state.label().c_str());
+    RCLCPP_INFO(get_logger(), "on shutdown is called from state %s.", state.label().c_str());
 
     // We return a success and hence invoke the transition to the next
     // step: "finalized".
