@@ -30,9 +30,12 @@ LoggerConfig::LoggerConfig(rclcpp::NodeOptions options)
 : Node("logger_config", options)
 {
   srv_ = create_service<logging_demo::srv::ConfigLogger>(
-    "config_logger", std::bind(
-      &LoggerConfig::handle_logger_config_request,
-      this, std::placeholders::_1, std::placeholders::_2));
+    "config_logger", [this](
+      const std::shared_ptr<logging_demo::srv::ConfigLogger::Request> request,
+      std::shared_ptr<logging_demo::srv::ConfigLogger::Response> response) {
+      return this->handle_logger_config_request(request, response);
+    }
+  );
 }
 
 void
