@@ -33,8 +33,10 @@ LoggerUsage::LoggerUsage(rclcpp::NodeOptions options)
 : Node("logger_usage_demo", options), count_(0)
 {
   pub_ = create_publisher<std_msgs::msg::String>("logging_demo_count", 10);
-  timer_ = create_wall_timer(500ms, std::bind(&LoggerUsage::on_timer, this));
-  debug_function_to_evaluate_ = std::bind(is_divisor_of_twelve, std::cref(count_), get_logger());
+  timer_ = create_wall_timer(500ms, [this]() {return this->on_timer();});
+  debug_function_to_evaluate_ = [this]() {
+      return is_divisor_of_twelve(std::cref(this->count_), this->get_logger());
+    };
 
   // After 10 iterations the severity will be set to DEBUG.
   auto on_one_shot_timer =
