@@ -36,50 +36,48 @@ public:
   {
     using namespace std::placeholders;
 
+    auto handle_goal = [this](
+      const rclcpp_action::GoalUUID & uuid,
+      std::shared_ptr<const Fibonacci::Goal> goal)
+      {
+        (void)uuid;
+        RCLCPP_INFO(this->get_logger(), "Received goal request with order %d", goal->order);
+        // The Fibonacci action uses int32 for the return of sequences, which means it can only hold
+        // 2^31-1 (2147483647) before wrapping negative in two's complement. Based on empirical
+        // tests, that means that an order of > 46 will cause wrapping, so we don't allow that here.
+        if (goal->order > 46) {
+          return rclcpp_action::GoalResponse::REJECT;
+        }
+        return rclcpp_action::GoalResponse::ACCEPT_AND_EXECUTE;
+      };
+
+    auto handle_cancel = [this](
+      const std::shared_ptr<GoalHandleFibonacci> goal_handle)
+      {
+        RCLCPP_INFO(this->get_logger(), "Received request to cancel goal");
+        (void)goal_handle;
+        return rclcpp_action::CancelResponse::ACCEPT;
+      };
+
+    auto handle_accepted = [this](
+      const std::shared_ptr<GoalHandleFibonacci> goal_handle)
+      {
+        // this needs to return quickly to avoid blocking the executor,
+        // so we declare a lambda function to be called inside a new thread
+        auto execute_in_thread = [this, goal_handle]() {return this->execute(goal_handle);};
+        std::thread{execute_in_thread}.detach();
+      };
+
     this->action_server_ = rclcpp_action::create_server<Fibonacci>(
-      this->get_node_base_interface(),
-      this->get_node_clock_interface(),
-      this->get_node_logging_interface(),
-      this->get_node_waitables_interface(),
+      this,
       "fibonacci",
-      std::bind(&FibonacciActionServer::handle_goal, this, _1, _2),
-      std::bind(&FibonacciActionServer::handle_cancel, this, _1),
-      std::bind(&FibonacciActionServer::handle_accepted, this, _1));
+      handle_goal,
+      handle_cancel,
+      handle_accepted);
   }
 
 private:
   rclcpp_action::Server<Fibonacci>::SharedPtr action_server_;
-
-  ACTION_TUTORIALS_CPP_LOCAL
-  rclcpp_action::GoalResponse handle_goal(
-    const rclcpp_action::GoalUUID & uuid,
-    std::shared_ptr<const Fibonacci::Goal> goal)
-  {
-    RCLCPP_INFO(this->get_logger(), "Received goal request with order %d", goal->order);
-    (void)uuid;
-    // Let's reject sequences that are over 9000
-    if (goal->order > 9000) {
-      return rclcpp_action::GoalResponse::REJECT;
-    }
-    return rclcpp_action::GoalResponse::ACCEPT_AND_EXECUTE;
-  }
-
-  ACTION_TUTORIALS_CPP_LOCAL
-  rclcpp_action::CancelResponse handle_cancel(
-    const std::shared_ptr<GoalHandleFibonacci> goal_handle)
-  {
-    RCLCPP_INFO(this->get_logger(), "Received request to cancel goal");
-    (void)goal_handle;
-    return rclcpp_action::CancelResponse::ACCEPT;
-  }
-
-  ACTION_TUTORIALS_CPP_LOCAL
-  void handle_accepted(const std::shared_ptr<GoalHandleFibonacci> goal_handle)
-  {
-    using namespace std::placeholders;
-    // this needs to return quickly to avoid blocking the executor, so spin up a new thread
-    std::thread{std::bind(&FibonacciActionServer::execute, this, _1), goal_handle}.detach();
-  }
 
   ACTION_TUTORIALS_CPP_LOCAL
   void execute(const std::shared_ptr<GoalHandleFibonacci> goal_handle)
