@@ -13,8 +13,6 @@
 // limitations under the License.
 
 #include <chrono>
-#include <cstdio>
-#include <memory>
 
 #include "rclcpp/rclcpp.hpp"
 #include "rclcpp_components/register_node_macro.hpp"
@@ -26,24 +24,23 @@ using namespace std::chrono_literals;
 namespace demo_nodes_cpp
 {
 
-class OneOffTimerNode : public rclcpp::Node
+class OneOffTimerNode final : public rclcpp::Node
 {
 public:
   DEMO_NODES_CPP_PUBLIC
   explicit OneOffTimerNode(const rclcpp::NodeOptions & options)
-  : Node("one_off_timer", options), count(0)
+  : Node("one_off_timer", options), count_(0)
   {
-    setvbuf(stdout, NULL, _IONBF, BUFSIZ);
-    periodic_timer = this->create_wall_timer(
+    periodic_timer_ = this->create_wall_timer(
       2s,
       [this]() {
         RCLCPP_INFO(this->get_logger(), "in periodic_timer callback");
-        if (this->count++ % 3 == 0) {
+        if (this->count_++ % 3 == 0) {
           RCLCPP_INFO(this->get_logger(), "  resetting one off timer");
-          this->one_off_timer = this->create_wall_timer(
+          this->one_off_timer_ = this->create_wall_timer(
             1s, [this]() {
               RCLCPP_INFO(this->get_logger(), "in one_off_timer callback");
-              this->one_off_timer->cancel();
+              this->one_off_timer_->cancel();
             });
         } else {
           RCLCPP_INFO(this->get_logger(), "  not resetting one off timer");
@@ -51,9 +48,10 @@ public:
       });
   }
 
-  rclcpp::TimerBase::SharedPtr periodic_timer;
-  rclcpp::TimerBase::SharedPtr one_off_timer;
-  size_t count;
+private:
+  rclcpp::TimerBase::SharedPtr periodic_timer_;
+  rclcpp::TimerBase::SharedPtr one_off_timer_;
+  size_t count_;
 };
 
 }  // namespace demo_nodes_cpp
