@@ -12,10 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <chrono>
 #include <iomanip>
 #include <iostream>
 #include <memory>
 #include <string>
+#include <thread>
 
 #include "std_msgs/msg/string.hpp"
 #include "rcutils/cmdline_parser.h"
@@ -33,7 +35,9 @@ static constexpr char OPTION_DEADLINE_PERIOD[] = "--deadline";
 static constexpr char OPTION_LIVELINESS_KIND[] = "--liveliness";
 static constexpr char OPTION_LEASE_DURATION[] = "--lease";
 
-void print_usage(const char * progname)
+static bool running = true;
+
+static void print_usage(const char * progname)
 {
   std::cout << progname << " [OPTIONS]" << std::endl <<
     std::endl << "Options when starting the demo:" << std::endl <<
@@ -61,39 +65,10 @@ void print_usage(const char * progname)
     std::endl;
 }
 
-class PublisherCommandHandler : public CommandGetter
+static void quit(void)
 {
-public:
-  PublisherCommandHandler(
-    rclcpp::executors::SingleThreadedExecutor & exec,
-    std::weak_ptr<Talker> publisher)
-  : exec_(exec), publisher_(publisher) {}
-
-  void handle_cmd(const char command) const override
-  {
-    const char cmd = tolower(command);
-    if (cmd == 'x') {
-      // signal program exit
-      exec_.cancel();
-      std::cout << "exiting the demo..." << std::endl;
-    } else if (auto publisher = publisher_.lock()) {
-      if (cmd == 'p') {
-        // manually assert liveliness of publisher
-        publisher->assert_publisher_liveliness();
-      } else if (cmd == 's') {
-        // toggle publishing of messages
-        publisher->toggle_publish();
-      } else if (cmd == 'q') {
-        // print the qos settings
-        publisher->print_qos();
-      }
-    }
-  }
-
-private:
-  rclcpp::executors::SingleThreadedExecutor & exec_;
-  std::weak_ptr<Talker> publisher_;
-};
+  running = false;
+}
 
 int main(int argc, char * argv[])
 {
@@ -153,16 +128,48 @@ int main(int argc, char * argv[])
         event.total_count, event.total_count_change);
     };
 
-  PublisherCommandHandler cmd_handler(exec, talker);
-
   talker->initialize();
   talker->print_qos();
 
-  cmd_handler.start();
+  install_ctrl_handler(quit);
+
   exec.add_node(talker);
-  exec.spin();
+  std::thread t([&exec] {
+      while (running) {
+        exec.spin_some();
+        std::this_thread::sleep_for(std::chrono::milliseconds(20));
+      }
+    });
+
+  KeyboardReader input;
+  char c;
+  while (running) {
+    try {
+      c = input.readOne();
+    } catch (const std::runtime_error &) {
+      running = false;
+      break;
+    }
+
+    switch (tolower(c)) {
+      case 'x':
+        running = false;
+        break;
+      case 'p':
+        talker->assert_publisher_liveliness();
+        break;
+      case 's':
+        talker->toggle_publish();
+        break;
+      case 'q':
+        talker->print_qos();
+        break;
+    }
+  }
+
+  t.join();
+
   exec.remove_node(talker);
-  cmd_handler.stop();
 
   rclcpp::shutdown();
 
