@@ -15,8 +15,8 @@
 #ifndef UTILS_HPP_
 #define UTILS_HPP_
 
-#include <atomic>
-#include <thread>
+#include <functional>
+#include <memory>
 
 #include "rclcpp/qos.hpp"
 #include "rmw/types.h"
@@ -29,30 +29,22 @@ rmw_time_to_seconds(const rmw_time_t & time);
 void
 print_qos(const rclcpp::QoS & qos);
 
-class CommandGetter
+void
+install_ctrl_handler(std::function<void(void)> ctrl_handler);
+
+class KeyboardReader final
 {
 public:
-  /// Whether or not the command getter is currently expecting listening for user inputs.
-  bool is_active() const;
+  KeyboardReader();
 
-  /// Start listening for user inputs (character key presses).
-  void start();
+  char readOne();
 
-  /// Stop listening for user inputs.
-  void stop();
-
-  /// Function prototype for handling user inputs.
-  virtual void handle_cmd(const char cmd) const = 0;
-
-  /// Helper function with event loop for the command getting thread.
-  void operator()() const;
+  ~KeyboardReader();
 
 private:
-  /// Get a key press from running terminal
-  char getch() const;
+  class KeyboardReaderImpl;
 
-  std::thread thread_;
-  std::atomic<bool> run_;
+  std::unique_ptr<KeyboardReaderImpl> pimpl_;
 };
 
 #endif  // UTILS_HPP_
