@@ -20,6 +20,9 @@ from quality_of_service_demo_py.common_nodes import Talker
 
 import rclpy
 from rclpy.duration import Duration
+from rclpy.event_handler import PublisherEventCallbacks
+from rclpy.event_handler import SubscriptionEventCallbacks
+from rclpy.event_handler import UnsupportedEventTypeError
 from rclpy.executors import ExternalShutdownException
 from rclpy.executors import SingleThreadedExecutor
 from rclpy.logging import get_logger
@@ -27,9 +30,6 @@ from rclpy.qos import QoSDurabilityPolicy
 from rclpy.qos import QoSLivelinessPolicy
 from rclpy.qos import QoSProfile
 from rclpy.qos import QoSReliabilityPolicy
-from rclpy.qos_event import PublisherEventCallbacks
-from rclpy.qos_event import SubscriptionEventCallbacks
-from rclpy.qos_event import UnsupportedEventTypeError
 
 
 def get_parser():
@@ -56,8 +56,8 @@ def main(args=None):
     if qos_policy_name == 'durability':
         print(
             'Durability incompatibility selected.\n'
-            'Incompatibility condition: publisher durability kind <'
-            'subscripition durability kind.\n'
+            'Incompatibility condition: publisher durability kind < '
+            'subscription durability kind.\n'
             'Setting publisher durability to: VOLATILE\n'
             'Setting subscription durability to: TRANSIENT_LOCAL\n'
         )
@@ -110,17 +110,30 @@ def main(args=None):
     else:
         print('{name} not recognised.'.format(name=qos_policy_name))
         parser.print_help()
-        return 0
+        return 1
 
     # Initialization and configuration
     rclpy.init(args=args)
     topic = 'incompatible_qos_chatter'
     num_msgs = 5
 
-    publisher_callbacks = PublisherEventCallbacks(
-        incompatible_qos=lambda event: get_logger('Talker').info(str(event)))
+    def sub_incompatible_qos_event(event):
+        count = event.total_count
+        delta = event.total_count_change
+        policy = event.last_policy_kind
+        get_logger('listener').info(
+            f'Requested incompatible qos - total {count} delta {delta} last_policy_kind: {policy}')
+
+    def pub_incompatible_qos_event(event):
+        count = event.total_count
+        delta = event.total_count_change
+        policy = event.last_policy_kind
+        get_logger('talker').info(
+            f'Offered incompatible qos - total {count} delta {delta} last_policy_kind: {policy}')
+
+    publisher_callbacks = PublisherEventCallbacks(incompatible_qos=pub_incompatible_qos_event)
     subscription_callbacks = SubscriptionEventCallbacks(
-        incompatible_qos=lambda event: get_logger('Listener').info(str(event)))
+        incompatible_qos=sub_incompatible_qos_event)
 
     try:
         talker = Talker(
@@ -141,10 +154,8 @@ def main(args=None):
     try:
         while talker.publish_count < num_msgs:
             executor.spin_once()
-    except KeyboardInterrupt:
+    except (KeyboardInterrupt, ExternalShutdownException):
         pass
-    except ExternalShutdownException:
-        return 1
     finally:
         rclpy.try_shutdown()
 
